// SDKs/twilio.com/twilio_messaging_v1/1.37.3/typescript/src/sdk/models/shared/messagingv1servicealphasender.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";



export class MessagingV1ServiceAlphaSender extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @SpeakeasyMetadata({ data: "json, name=alpha_sender" })
  alphaSender?: string;

  @SpeakeasyMetadata({ data: "json, name=capabilities" })
  capabilities?: string[];

  @SpeakeasyMetadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @SpeakeasyMetadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @SpeakeasyMetadata({ data: "json, name=service_sid" })
  serviceSid?: string;

  @SpeakeasyMetadata({ data: "json, name=sid" })
  sid?: string;

  @SpeakeasyMetadata({ data: "json, name=url" })
  url?: string;
}